requirements.txt
lexica-api==1.5.7
numpy==1.26.4
opencv-python
aiofiles==23.2.1
aiohttp==3.9.1
APScheduler==3.10.4
alphabet-detector==0.0.7
asyncio==3.4.3
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
dateparser==1.2.0
emoji==2.9.0
emoji-country-flag==1.3.2
ffmpeg-python
future==1.0.0
gpytranslate==1.5.1
lxml==5.1.0
html2text==2020.1.16
httpx[http2]
humanize==4.9.0
jikanpy==4.3.2
markdown2==2.4.13
motor==3.3.2
nekos.py==1.1.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
psutil==5.9.8
pyjokes
psycopg2-binary==2.9.9
pynewtonmath==1.0.1
pymongo[srv]==4.6.2
pyrate-limiter==3.1.1
pytz==2024.1
python-arq==6.0.5
python-dotenv
pyrogram==2.0.106
python-telegram-bot==20.8
regex==2023.12.25
requests==2.31.0
shortuuid==1.0.11
SQLAlchemy==2.0.27
speedtest-cli==2.1.3
telegraph==2.2.0
telethon==1.34.0
tgcrypto==1.2.5
Unidecode==1.3.8
youtube_dl
youtube_search_python
youtube_search
search_engine_parser
wget
tswift
regex

